+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Change `window.set_outer_position()` to take an `impl Into<glam::IVec2>`.
   * Change `window.set_inner_size()` to take an `impl Into<glam::UVec2>`.
   * Bump maximum `tch` version to `0.8.x`.
+  * Use `glam` version `0.20` or `0.21`.
   * Update internal dependencies.
 
 # 0.12.3 - 2022-05-22